backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
networkx==3.4.2
groq==0.15.0
python-dotenv==1.0.1
pydantic==2.10.3
